# Optimize exception cleanup