@ConradIrwin 2015-04-18
Thanks for this. Neat trick with memoizing the clean object in the recursion avoider :).

There's probably plenty of low-hanging performance fruit if you're feeling keen ;)