@ConradIrwin 2015-04-18
(released in v2.8.5)